% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.object.R, R/pi.object1.R
\name{pi.object}
\alias{pi.object}
\title{pi.object}
\usage{
pi.object(df)

pi.object(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Object suitable for IBM Personality Insights pull

Object suitable for IBM Personality Insights pull
}
\description{
pi.object

pi.object1
}
