% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomLetter.R
\name{randomLetter}
\alias{randomLetter}
\title{Random Letter Generator}
\usage{
randomLetter(n, letter_probs = rep(1/26, 26), wreplacement = TRUE)
}
\arguments{
\item{n}{number of desired random letters}

\item{letter_probs}{a vector of probabilities for each of the 26 letters (optional)}

\item{wreplacement}{if you want multiple letters, logical TRUE/FALSE for sampling with replacement}
}
\value{
a random letter based on probabilities
}
\description{
Random Letter Generator
}
\examples{
randomLetter(1) #selecct 1 random letter
randomLetter(2) #select 2 random letters
}
